Add a possibility to attach files (#1)
Added a possibility to upload files now along the paste as attachments.
Due to this feature you can easily now upload smaller config files to paste.me to securely share it with others. This is not intended to store files long-term, it is really just a way to quickly move files between instances without having to copy/paste it or send it in an unsecure manner.
Tests added too!